Forgot to merge the two branches

+++ func/resultdelete.func.php
@@ -7,16 +7,16 @@ if(!isset($_SESSION["is_admin"])){
 	header('Location: ' . $_SERVER['HTTP_REFERER']);
 } else {
 	mysql_connect($db_host, $db_user, $db_password) or die(mysql_error()); 
-	mysql_select_db($_SESSION['dbext']) or die(mysql_error()); 
+	mysql_select_db($db_name) or die(mysql_error()); 
 	if(isset($_GET['uuid'])){
 		$uuid = mysql_real_escape_string($_GET['uuid']);
 		if($uuid=="all"){
-			$query = mysql_query("TRUNCATE TABLE results");
+			$query = mysql_query("DELETE FROM results WHERE owner = '".$_SESSION['dbext']."'");
 		} else if(!is_int($uuid) && isset($_GET['house'])){
 			$house = $_GET['house'];
-			$query = mysql_query("DELETE FROM results WHERE quizuuid='$uuid' AND house='$house';");
+			$query = mysql_query("DELETE FROM results WHERE quizuuid='$uuid' AND house='$house' AND owner='".$_SESSION['dbext']."';");
 		} else {
-			$query = mysql_query("DELETE FROM results WHERE id='$uuid'");
+			$query = mysql_query("DELETE FROM results WHERE id='$uuid' AND owner='".$_SESSION['dbext']."'");
 		}
 		mysql_close();
 		header('Location: ' . $_SERVER['HTTP_REFERER']);
@@ -27,7 +27,7 @@ if(!isset($_SESSION["is_admin"])){
 		} else {
 			$uuid = mysql_real_escape_string($_POST['uuid']);
 			$house = $_POST['house'];
-			$query = mysql_query("DELETE FROM results WHERE quizuuid='$uuid' AND house='$house';");
+			$query = mysql_query("DELETE FROM results WHERE quizuuid='$uuid' AND house='$house' AND owner='".$_SESSION['dbext']."';");
 			header('Location: ' . $_SERVER['HTTP_REFERER']);
 		}
 	}
